[Refactor] seperated models to each model
