+++ actoon/views/effectview.py
@@ -2,20 +2,20 @@ from django.shortcuts import get_list_or_404, get_object_or_404
 from rest_framework import viewsets, permissions
 from rest_framework.response import Response
 
+import actoon.models.effectmodel
 import actoon.serializers.effectserializer
-from actoon import models as actoon_model
 
 
 class EffectView(viewsets.ModelViewSet):
-    queryset = actoon_model.Effect.objects.all()
+    queryset = actoon.models.effectmodel.Effect.objects.all()
     serializer_class = actoon.serializers.effectserializer.EffectSerializer
     permission_classes = (permissions.IsAuthenticated,)
 
     def get_queryset(self, pk=None):
         if pk is not None:
-            queryset = actoon_model.Effect.objects.all(pk=pk)
+            queryset = actoon.models.effectmodel.Effect.objects.all(pk=pk)
         else:
-            queryset = actoon_model.Effect.objects.all()
+            queryset = actoon.models.effectmodel.Effect.objects.all()
 
         return queryset
